requirements.txt
python=3.8.16
tensorflow-gpu=2.6.0
cudatoolkit=11.3.1
cudnn=8.2.1
numpy=1.20.3
pandas=2.0.3
seaborn=0.12.2
matplotlib=3.7.2
scikit-learn=1.3.1
tqdm=4.66.1